Removing FigurePanel support.
This should not have been added to Texture, as there has not been consensus
in the need of panels at all.
