--- data/blank-figure-package/figure.xml
@@ -1,45 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE article PUBLIC "-//NLM//DTD JATS (Z39.96) Journal Archiving and Interchange DTD v1.2 20190208//EN"  "JATS-archivearticle1.dtd">
-<article xmlns:xlink="http://www.w3.org/1999/xlink" xmlns:ali="http://www.niso.org/schemas/ali/1.0" article-type="figure-package">
-  <front>
-    <article-meta>
-      <title-group>
-        <article-title />
-      </title-group>
-      <abstract />
-    </article-meta>
-  </front>
-  <body>
-    <fig-group id="fig1">
-      <fig id="fig1a">
-        <label>Figure 1A</label>
-        <caption id="fig1a-caption">
-          <title></title>
-          <p id="fig1b-caption-p1"></p>
-        </caption>
-        <kwd-group>
-          <label>Exp system</label>
-          <kwd />
-        </kwd-group>
-        <kwd-group>
-          <label>Method</label>
-          <kwd />
-        </kwd-group>
-        <kwd-group>
-          <label>Measured variable</label>
-          <kwd />
-        </kwd-group>
-        <kwd-group>
-          <label>Perturbations</label>
-          <kwd />
-        </kwd-group>
-        <kwd-group>
-          <label>Disease</label>
-          <kwd />
-        </kwd-group>
-        <graphic id="fig1b-graphic" mime-subtype="jpeg" mimetype="image" xlink:href="placeholder.png" />
-      </fig>
-    </fig-group>
-  </body>
-  <back />
-</article>
