--- data/blank-figure-package/manifest.xml
@@ -1,8 +1,0 @@
-<dar>
-  <documents>
-    <document id="manuscript" type="article" path="figure.xml" />
-  </documents>
-  <assets>
-    <asset id="placeholder" type="image/png" path="placeholder.png"/>
-  </assets>
-</dar>
