--- data/figure-package/manifest.xml
@@ -1,10 +1,0 @@
-<dar>
-  <documents>
-    <document id="manuscript" type="article" path="figure.xml" />
-  </documents>
-  <assets>
-    <asset id="fig1a" type="image/jpg" path="fig1a.jpg"/>
-    <asset id="fig1a-source-data" type="text/csv" path="fig1a-source-data.csv"/>
-    <asset id="fig1b" type="image/jpg" path="fig1b.jpg"/>
-  </assets>
-</dar>
