+++ src/article/TextureJATS.rng
@@ -54,6 +54,7 @@
   <s:not-implemented name="equation-count"/>
   <s:not-implemented name="etal"/>
   <s:not-implemented name="fig-count"/>
+  <s:not-implemented name="fig-group"/>
   <s:not-implemented name="floats-group"/>
   <s:not-implemented name="front-stub"/>
   <s:not-implemented name="funding-statement"/>
@@ -532,7 +533,6 @@
       <ref name="boxed-text"/>
       <ref name="chem-struct-wrap"/>
       <ref name="fig"/>
-      <ref name="fig-group"/>
       <ref name="table-wrap"/>
       <ref name="disp-formula"/>
       <ref name="disp-formula-group"/>
@@ -753,25 +753,10 @@
       <optional>
         <ref name="caption"/>
       </optional>
-      <zeroOrMore>
-        <ref name="kwd-group"/>
-      </zeroOrMore>
       <ref name="graphic"/>
       <optional>
         <ref name="permissions"/>
       </optional>
-    </element>
-  </define>
-
-  <define name="fig-group">
-    <element name="fig-group">
-      <ref name="fig-group-attlist"/>
-      <optional>
-        <ref name="label"/>
-      </optional>
-      <oneOrMore>
-        <ref name="fig"/>
-      </oneOrMore>
     </element>
   </define>
