+++ src/article/commands/index.js
@@ -9,8 +9,6 @@ export { default as EditAuthorCommand } from './EditAuthorCommand'
 export { default as EditEntityCommand } from './EditEntityCommand'
 export { default as EditMetadataCommand } from './EditMetadataCommand'
 export { default as EditReferenceCommand } from './EditReferenceCommand'
-export * from './FigureMetadataCommands'
-export * from './FigurePanelCommands'
 export { default as IncreaseHeadingLevelCommand } from './IncreaseHeadingLevelCommand'
 export { default as InsertBlockFormulaCommand } from './InsertBlockFormulaCommand'
 export { default as InsertBlockQuoteCommand } from './InsertBlockQuoteCommand'
