+++ src/article/components/FigureComponent.js
@@ -1,110 +1,74 @@
 import { $$ } from 'substance'
 import { NodeComponent } from '../../kit'
+import { PREVIEW_MODE } from '../ArticleConstants'
+import PreviewComponent from './PreviewComponent'
+import LabelComponent from './LabelComponent'
+import { getLabel } from '../shared/nodeHelpers'
 
 export default class FigureComponent extends NodeComponent {
-  /*
-    Note: in the Manuscript View only one figure panel is shown at time.
-  */
   render () {
-    let mode = this._getMode()
-    let node = this.props.node
-    let panels = node.panels
-
-    let el = $$('div').addClass('sc-figure').addClass(`sm-${mode}`).attr('data-id', node.id)
-    if (panels.length > 0) {
-      let content = this._renderCarousel(panels)
-      el.append(content)
+    const mode = this._getMode()
+    // different rendering when rendered as preview or in metadata view
+    if (mode === PREVIEW_MODE) {
+      return this._renderPreviewVersion()
+    } else {
+      return this._renderManuscriptVersion()
     }
+  }
+
+  _getClassNames () {
+    return `sc-figure`
+  }
+
+  _renderManuscriptVersion () {
+    const mode = this._getMode()
+    const node = this.props.node
+    const SectionLabel = this.getComponent('section-label')
+
+    let el = $$('div')
+      .addClass(this._getClassNames())
+      .attr('data-id', node.id)
+      .addClass(`sm-${mode}`)
+
+    el.append(
+      $$(SectionLabel, { label: 'label-label' }),
+      $$(LabelComponent, { node }),
+      // no label for the graphic
+      this._renderContent(),
+      $$(SectionLabel, { label: 'title-label' }),
+      this._renderValue('title', { placeholder: this.getLabel('title-placeholder') }).addClass('se-title'),
+      $$(SectionLabel, { label: 'legend-label' }),
+      this._renderValue('legend', { placeholder: this.getLabel('legend-placeholder') }).addClass('se-legend')
+    )
+
     return el
   }
 
-  _renderCarousel (panels) {
-    if (panels.length === 1) {
-      return this._renderCurrentPanel()
-    } else {
-      return $$('div').addClass('se-carousel').append(
-        this._renderNavigation(),
-        $$('div').addClass('se-current-panel').append(
-          this._renderCurrentPanel()
-        )
-      )
-    }
+  _renderContent () {
+    return this._renderValue('content').addClass('se-content')
   }
 
-  _renderCurrentPanel () {
-    let panel = this._getCurrentPanel()
-    let PanelComponent = this.getComponent(panel.type)
-    return $$(PanelComponent, {
-      node: panel,
-      mode: this.props.mode
-    }).ref(panel.id)
-  }
-
-  _renderNavigation () {
+  _renderPreviewVersion () {
     const node = this.props.node
-    const panels = node.getPanels()
-    const numberOfPanels = panels.length
-    const currentIndex = this._getCurrentPanelIndex() + 1
-    const currentPosition = currentIndex + ' / ' + numberOfPanels
-    const leftControl = $$('div').addClass('se-control sm-previous').append(
-      this._renderIcon('left-control')
-    )
-    if (currentIndex > 1) {
-      leftControl.on('click', this._onSwitchPanel.bind(this, 'left'))
-    } else {
-      leftControl.addClass('sm-disabled')
+    // TODO: We could return the PreviewComponent directly.
+    // However this yields an error we need to investigate.
+    let thumbnail
+    let content = node.resolve('content')
+    if (content.type === 'graphic') {
+      let ContentComponent = this.getComponent(content.type)
+      thumbnail = $$(ContentComponent, {
+        node: content
+      })
     }
-    const rightControl = $$('div').addClass('se-control sm-next').append(this._renderIcon('right-control'))
-    if (currentIndex < numberOfPanels) {
-      rightControl.on('click', this._onSwitchPanel.bind(this, 'right'))
-    } else {
-      rightControl.addClass('sm-disabled')
-    }
-    return $$('div').addClass('se-navigation').append(
-      $$('div').addClass('se-current-position').append(currentPosition),
-      $$('div').addClass('se-controls').append(
-        leftControl,
-        rightControl
-      )
-    )
+    // TODO: PreviewComponent should work with a model
+    return $$('div').append($$(PreviewComponent, {
+      id: node.id,
+      thumbnail,
+      label: getLabel(node)
+    })).addClass('sc-figure-panel').attr('data-id', node.id)
   }
 
   _getMode () {
     return this.props.mode || 'manuscript'
-  }
-
-  _getCurrentPanel () {
-    let node = this.props.node
-    let doc = node.getDocument()
-    let currentPanelIndex = this._getCurrentPanelIndex()
-    let ids = node.panels
-    return doc.get(ids[currentPanelIndex])
-  }
-
-  _getCurrentPanelIndex () {
-    let node = this.props.node
-    let state = node.state
-    let panels = node.panels
-    let currentPanelIndex = 0
-    if (state) {
-      currentPanelIndex = state.currentPanelIndex
-    }
-    // FIXME: state is corrupt
-    if (currentPanelIndex < 0 || currentPanelIndex >= panels.length) {
-      console.error('figurePanel.state.currentPanelIndex is corrupt')
-      state.currentPanelIndex = currentPanelIndex = 0
-    }
-    return currentPanelIndex
-  }
-
-  _onSwitchPanel (direction) {
-    let currentIndex = this._getCurrentPanelIndex()
-    this.context.api.switchFigurePanel(this.props.node, direction === 'left' ? --currentIndex : ++currentIndex)
-  }
-
-  _renderIcon (iconName) {
-    return $$('div').addClass('se-icon').append(
-      this.context.iconProvider.renderIcon(iconName)
-    )
   }
 }
