+++ src/article/components/index.js
@@ -14,9 +14,6 @@ export { default as EditorPanel } from './EditorPanel'
 export { default as ExternalLinkComponent } from './ExternalLinkComponent'
 export { default as ExternalLinkEditor } from './ExternalLinkEditor'
 export { default as FigureComponent } from './FigureComponent'
-export { default as FigureMetadataComponent } from './FigureMetadataComponent'
-export { default as FigurePanelComponent } from './FigurePanelComponent'
-export { default as FigurePanelComponentWithMetadata } from './FigurePanelComponentWithMetadata'
 export { default as FileUploadComponent } from './FileUploadComponent'
 export { default as FootnoteComponent } from './FootnoteComponent'
 export { default as FootnoteEditor } from './FootnoteEditor'
@@ -25,7 +22,6 @@ export { default as HeadingComponent } from './HeadingComponent'
 export { default as InlineFormulaComponent } from './InlineFormulaComponent'
 export { default as InlineFormulaEditor } from './InlineFormulaEditor'
 export { default as InlineGraphicComponent } from './InlineGraphicComponent'
-export { default as InsertFigurePanelTool } from './InsertFigurePanelTool'
 export { default as InsertFigureTool } from './InsertFigureTool'
 export { default as InsertInlineGraphicTool } from './InsertInlineGraphicTool'
 export { default as InsertTableTool } from './InsertTableTool'
@@ -37,16 +33,13 @@ export { default as ListItemComponent } from './ListItemComponent'
 export { default as ManuscriptComponent } from './ManuscriptComponent'
 export { default as ManuscriptSection } from './ManuscriptSection'
 export { default as ManuscriptTOC } from './ManuscriptTOC'
-export { default as MetadataFieldComponent } from './MetadataFieldComponent'
 export { default as ModelPreviewComponent } from './ModelPreviewComponent'
-export { default as OpenFigurePanelImageTool } from './OpenFigurePanelImageTool'
 export { default as ParagraphComponent } from './ParagraphComponent'
 export { default as PreviewComponent } from './PreviewComponent'
 export { default as QueryComponent } from './QueryComponent'
 export { default as ReferenceComponent } from './ReferenceComponent'
 export { default as ReferenceListComponent } from './ReferenceListComponent'
 export { default as ReferenceUploadComponent } from './ReferenceUploadComponent'
-export { default as ReplaceFigurePanelTool } from './ReplaceFigurePanelTool'
 export { default as ReplaceSupplementaryFileTool } from './ReplaceSupplementaryFileTool'
 export { default as SectionLabel } from './SectionLabel'
 export { default as SubscriptComponent } from './SubscriptComponent'
@@ -58,7 +51,6 @@ export { default as TableCellEditor } from './TableCellEditor'
 export { default as TableComponent } from './TableComponent'
 export { default as TableContextMenu } from './TableContextMenu'
 export { default as TableFigureComponent } from './TableFigureComponent'
-export { default as TableFigureComponentWithMetadata } from './TableFigureComponentWithMetadata'
 export { default as UnsupportedInlineNodeComponent } from './UnsupportedInlineNodeComponent'
 export { default as UnsupportedNodeComponent } from './UnsupportedNodeComponent'
 export { default as UploadSingleImageTool } from './UploadSingleImageTool'
