+++ src/article/nodes/index.js
@@ -11,11 +11,9 @@ export { default as Break } from './Break'
 export { default as ChapterRef } from './ChapterRef'
 export { default as ConferencePaperRef } from './ConferencePaperRef'
 export { default as CustomAbstract } from './CustomAbstract'
-export { default as MetadataField } from './MetadataField'
 export { default as DataPublicationRef } from './DataPublicationRef'
 export { default as ExternalLink } from './ExternalLink'
 export { default as Figure } from './Figure'
-export { default as FigurePanel } from './FigurePanel'
 export { default as Footnote } from './Footnote'
 export { default as Funder } from './Funder'
 export { default as Group } from './Group'
