+++ src/article/settings/DefaultSettings.js
@@ -14,7 +14,6 @@ export default {
   'data-publication-ref.authors': { required: true },
   'data-publication-ref.containerTitle': { required: true },
   'data-publication-ref.title': { required: true },
-  'figure-panel.title': { required: true },
   'funder.institution': { required: true },
   'group.name': { required: true },
   'journal-article-ref.authors': { required: true },
