+++ src/article/shared/FigureLabelGenerator.js
@@ -1,5 +1,4 @@
 import { last, isArray } from 'substance'
-import { LATIN_LETTERS_UPPER_CASE } from '../ArticleConstants'
 
 export default class FigureLabelGenerator {
   constructor (config = {}) {
@@ -98,23 +97,12 @@ export default class FigureLabelGenerator {
       return String(def[0].pos)
     } else {
       let figCounter = def[0].pos
-      // TODO: we should think about some way to make this configurable
-      return `${figCounter}${this._getPanelLabel(def)}`
+      return `${figCounter}`
     }
-  }
-
-  _getPanelLabel (def) {
-    let panelCounter = def[1].pos
-    return `${LATIN_LETTERS_UPPER_CASE[panelCounter - 1]}`
   }
 
   _getGroupCounter (first, last) {
-    // ATTENTION: assuming that first and last have the same level (according to our implementation)
-    if (first.length === 1) {
-      return `${this._getSingleCounter(first)}${this.config.to}${this._getSingleCounter(last)}`
-    } else {
-      return `${this._getSingleCounter(first)}${this.config.to}${this._getPanelLabel(last)}`
-    }
+    return `${this._getSingleCounter(first)}${this.config.to}${this._getSingleCounter(last)}`
   }
 
   _replaceAll (t, $) {
